Fill in more fields in OPDS entries (BL-8089)
UNKNOWN is a placeholder value to alert us to either data is isn't in
the books table at all or data that is not filled in for the given book.
This is just a temporary device: a real catalog would probably just omit
the field.  I think id, title, and updated are the only required fields.

The show field in the books table should be used to filter whether the
various links are actually provided.  This has not yet been implemented.